Refactor PyInterpreter to use normal vtables
I realized that we can deal with the dead vtable problem by...
introducing another indirection!  The resulting code is worse
(you have to do one more dereference to get to the vtable), but
the reduction in boilerplate is, IMO, worth it.

I did this refactor because I'm about to add a lot more methods
to PyInterpreter to handle expunging SymInt from TensorImpl.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 0798b1f9c47ef2faf48452eb5293f395f63aab01
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84388