Optimise the decomposition for `adaptive_avg_pool2d` wrt. TorchInductor
This fixes some part of the implementation that did not work with
TorchInductor (e.g. the indices in TorchInductor need to be `int64`s,
while in PyTorch we can have `int32`s).

It also brings up the performance of the kernel to similar numbers than
those of the lowering (benchmarks below).

ghstack-source-id: 982532547222679b30dec7c49fc98107b9825050
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84483